Labfiles/01-build-agent-portal-and-vscode/Python/requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=22.0.0
